requirements_integration.txt
pandas>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
boto3>=1.34.0
pyarrow>=15.0.0
python-dotenv>=1.0.0
